backend/requirements.txt
Django>=3.0.2,<3.1.0
djangorestframework>=3.11.0,<3.12.0
django-cors-headers
psycopg2-binary
djangorestframework_simplejwt
django[argon2]
argon2-cffi
django-ratelimit